+++ force_matching.py
@@ -189,20 +189,31 @@ if __name__ == "__main__":
         fm_loss.assign_crossval_sets()
         fm_loss.calc_matrices(Ucg, forcenames, coll_var_names=psinames, verbose=True)
 
-    rdg_alphas = np.logspace(-10, 8, 500)
-    fm_loss.solve(rdg_alphas)
-
     os.chdir(cg_savedir)
 
-    np.save("rdg_cstar.npy", fm_loss.coeff_star)
-    #np.save("rdg_cstar.npy", fm_loss.coeff_star)
+    if not os.path.exists("rdg_valid_mse.npy"):
+        rdg_alphas = np.logspace(-10, 8, 500)
+        fm_loss.solve(rdg_alphas)
 
-    rdg_idxs = [5, 50, 200, 300]
-    plot_Ucg_vs_alpha(rdg_idxs, fm_loss.alpha_star_idx, fm_loss.coeffs, rdg_alphas, Ucg, cv_r0_basis, "rdg_")
-    #plot_Xcoeff_vs_d(rdg_idxs, rdg_idx_star, rdg_coeffs, rdg_alphas, fm_loss.X, fm_loss.d, "rdg_")
+        np.save("rdg_cstar.npy", fm_loss.coeff_star)
+        np.save("rdg_coeffs.npy", fm_loss.coeffs)
+        np.save("rdg_train_mse.npy", fm_loss.train_mse)
+        np.save("rdg_valid_mse.npy", fm_loss.valid_mse)
 
-    iff.util.plot_train_test_mse(rdg_alphas, fm_loss.train_mse, fm_loss.valid_mse, 
-            xlabel=r"Regularization $\alpha$", 
-            ylabel="Mean squared error (MSE)", 
-            title="Ridge regression", prefix="ridge_")
+        rdg_idxs = [5, 50, 200, 300]
+        plot_Ucg_vs_alpha(rdg_idxs, fm_loss.alpha_star_idx, fm_loss.coeffs, rdg_alphas, Ucg, cv_r0_basis, "rdg_")
+        #plot_Xcoeff_vs_d(rdg_idxs, rdg_idx_star, rdg_coeffs, rdg_alphas, fm_loss.X, fm_loss.d, "rdg_")
+
+        iff.util.plot_train_test_mse(rdg_alphas, fm_loss.train_mse, fm_loss.valid_mse, 
+                xlabel=r"Regularization $\alpha$", 
+                ylabel="Mean squared error (MSE)", 
+                title="Ridge regression", prefix="ridge_")
+
+    if not os.path.exists("rdg_fixed_sigma_cstar.npy"):
+        f_mult_12, alphas, all_coeffs, tr_mse, vl_mse = iff.util.scan_with_fixed_sigma(fm_loss, Ucg, cv_r0_basis)
+        sigma_idx, alpha_idx = np.argwhere(vl_mse[:,:,0] == vl_mse[:,:,0].min())[0]
+        new_coeffs = np.concatenate([ np.array([f_mult_12[sigma_idx]]), all_coeffs[sigma_idx, alpha_idx]])
+        np.save("rdg_fixed_sigma_cstar.npy", new_coeffs)
+    else:
+        new_coeffs = np.load("rdg_fixed_sigma_cstar.npy")
